requirements.txt
django <=3.0.7
mysqlclient <=1.4.6
djangorestframework